split out theme from App.tsx
