+++ frontend/src/App.tsx
@@ -1,13 +1,8 @@
 //@ts-check
 import React, {useState} from "react";
-import axios from "axios";
-import Button from '@mui/material/Button';
-import RestartAltIcon from '@mui/icons-material/RestartAlt';
 import CssBaseline from "@mui/material/CssBaseline";
 import Box from '@mui/material/Box';
 import FormControl from '@mui/material/FormControl';
-import TextField from '@mui/material/TextField';
-import MenuItem from '@mui/material/MenuItem';
 import Switch from '@mui/material/Switch';
 import Stack from '@mui/material/Stack';
 import Typography from '@mui/material/Typography';
@@ -15,74 +10,11 @@ import Tabs from '@mui/material/Tabs';
 import Tab from '@mui/material/Tab';
 import AppBar from '@mui/material/AppBar';
 import Toolbar from "@mui/material/Toolbar";
-import { createTheme, ThemeProvider, PaletteMode} from '@mui/material';
-import Table from '@mui/material/Table';
-import TableBody from '@mui/material/TableBody';
-import TableCell from '@mui/material/TableCell';
-import TableContainer from '@mui/material/TableContainer';
-import TableHead from '@mui/material/TableHead';
-import TableRow from '@mui/material/TableRow';
-import Paper from '@mui/material/Paper';
-import { handleTimeInput } from "./utils/inputValidation";
 import Calculator from "./components/Calculator";
 import { TabPanel } from "./components/TabPanel";
+import { ThemeProvider, PaletteMode} from '@mui/material';
+import getTheme from "./styles/theme";
 
-const getTheme = (mode: PaletteMode) =>
-  createTheme({
-    palette: {
-      mode, // Dynamically set mode ('light' or 'dark')
-      background: {
-        default: mode === "dark" ? "#121212" : "#F5F5F5",
-        paper: mode === "dark" ? "#1E1E1E" : "#FFFFFF",
-      },
-      primary: {
-        main: mode === "dark" ? "#BB86FC" : "#6200EE",
-        contrastText: "#FFFFFF",
-      },
-      secondary: {
-        main: mode === "dark" ? "#03DAC6" : "#018786",
-        contrastText: mode === "dark" ? "#000000" : "#FFFFFF",
-      },
-      text: {
-        primary: mode === "dark" ? "#FFFFFF" : "#212121",
-        secondary: mode === "dark" ? "#B0B0B0" : "#757575",
-        disabled: mode === "dark" ? "#666666" : "#BDBDBD",
-      },
-      divider: mode === "dark" ? "#333333" : "#E0E0E0",
-    },
-    typography: {
-      fontFamily: "Roboto, Arial, sans-serif",
-      h1: { fontSize: "2.5rem", fontWeight: 600, color: mode === "dark" ? "#FFFFFF" : "#212121" },
-      h2: { fontSize: "2rem", fontWeight: 500, color: mode === "dark" ? "#FFFFFF" : "#212121" },
-      body1: { fontSize: "1rem", color: mode === "dark" ? "#E0E0E0" : "#424242" },
-      body2: { fontSize: "0.875rem", color: mode === "dark" ? "#B0B0B0" : "#757575" },
-    },
-    components: {
-      MuiButton: {
-        styleOverrides: {
-          root: {
-            borderRadius: "8px",
-            textTransform: "none",
-            "&:hover": {
-              backgroundColor: mode === "dark" ? "#985EFF" : "#a574fc",
-            },
-          },
-          contained: {
-            boxShadow: mode === "dark" ? "0px 4px 6px rgba(0, 0, 0, 0.2)" : "0px 2px 4px rgba(0, 0, 0, 0.1)",
-          },
-        },
-      },
-      MuiPaper: {
-        styleOverrides: {
-          root: {
-            backgroundColor: mode === "dark" ? "#1E1E1E" : "#FFFFFF",
-            borderRadius: "12px",
-            boxShadow: mode === "dark" ? "none" : "0px 2px 8px rgba(0, 0, 0, 0.1)",
-          },
-        },
-      },
-    },
-  });
 
 function App() {
   const [mode, setMode] = useState<PaletteMode>("light"); // 'light' or 'dark'
